Minor code changes to allow build using Microsoft Visual Studio 2015 Code changes to allow builds that do not use ASM_FAST_MUTEX and/or POPCNT_CAPABILITY Visual Studio projects, associated files and Microsft specific additional code in the msvcc directory

+++ aligner_swsse_ee_i16.cpp
@@ -635,7 +635,7 @@ TAlScore SwAligner::alignGatherEE16(int& flag, bool debug) {
 			if(cperTri_) {
 				size_t rc_mod = coli & cper_.lomask_;
 				assert_lt(rc_mod, cper_.per_);
-				int64_t row = -rc_mod-1;
+				int64_t row = -(int64_t)rc_mod-1;
 				int64_t row_mod = row;
 				int64_t row_div = 0;
 				size_t idx = coli >> cper_.perpow2_;
