+++ aligner_swsse_loc_u8.cpp
@@ -736,7 +736,7 @@ TAlScore SwAligner::alignGatherLoc8(int& flag, bool debug) {
 				// Checkpoint for triangular mini-fills
 				size_t rc_mod = coli & cper_.lomask_;
 				assert_lt(rc_mod, cper_.per_);
-				int64_t row = -rc_mod-1;
+				int64_t row = -(int64_t)rc_mod-1;
 				int64_t row_mod = row;
 				int64_t row_div = 0;
 				size_t idx = coli >> cper_.perpow2_;
